+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html). However, all releases before 1.0.0 have breaking changes
 between minor-version updates.
 
-## [2.4.0] - Unreleased
+## [2.4.1] - Unreleased
+
+### Fixed
+
+- List construction error when compiling all event considerations
+
+## [2.4.0] - 2023-11-19
 
 ### Fixed